requirements.txt
mysql-connector-python~=8.0.33
requests~=2.31.0
python-dotenv~=1.0.0
pandas~=2.0.2
Flask~=2.3.2
SQLAlchemy~=2.0.15
Jinja2~=3.1.2
decouple~=0.0.7